# assets/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger CodeCraft",
        "description": "API endpoints documentations for CodeCraft Project"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/users/me": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "description": "Get current user",
                "operationId": "get_current_user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "current user response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User Management"
                ],
                "description": "update user informations",
                "operationId": "update_user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "credentials",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update_user response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/users/my_progress": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "description": "Get current user progress",
                "operationId": "getCurrentUserProgress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "current user response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/UserProgress"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/auth/sign_out": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "description": "Sign out from current session",
                "operationId": "sign_out",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sign_out response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/organizations/create": {
            "post": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Create a new organization",
                "operationId": "createOrganization",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create organization response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/Organization"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/challenges/create": {
            "post": {
                "tags": [
                    "Challenges Management"
                ],
                "description": "Creates a new challenge",
                "operationId": "createChallenge",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "create new challenge body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "topic": {
                                    "type": "string"
                                },
                                "level": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "is_public": {
                                    "type": "boolean"
                                }
                            },
                            "required": [
                                "name"
                            ]
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create challenge response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/Challenge"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/jobposts/create": {
            "post": {
                "tags": [
                    "Job Posting"
                ],
                "description": "Create a new Job Post",
                "operationId": "createJobPost",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create job post response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/JobPost"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/events/create": {
            "post": {
                "tags": [
                    "Events Management"
                ],
                "description": "Create a new Event",
                "operationId": "createEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "is_public": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "start_at": {
                                    "type": "string"
                                },
                                "end_at": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create event response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/Event"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/upload/image": {
            "post": {
                "tags": [
                    "Upload Management"
                ],
                "description": "Upload image",
                "operationId": "uploadImage",
                "produces": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "image",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "upload image response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "string",
                                    "example":"/public/images/0001020759_03052024.png"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
        "UserProgress": {
            "type": "object",
            "properties": {
                "correct_easy_submissions": {
                    "type": "integer"
                },
                "total_easy_submissions": {
                    "type": "integer"
                },
                "correct_medium_submissions": {
                    "type": "integer"
                },
                "total_medium_submissions": {
                    "type": "integer"
                },
                "correct_hard_submissions": {
                    "type": "integer"
                },
                "total_hard_submissions": {
                    "type": "integer"
                },
                "total_correct_submissions": {
                    "type": "integer"
                },
                "total_submissions": {
                    "type": "integer"
                }
            }
        },
        "Organization": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/User"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
        "Challenge": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
        "JobPost": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
...
        "Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "start_at": {
                    "type": "integer"
                },
                "end_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "organization_id": {
                    "type": "integer"
                },
                "organization": {
                    "$ref": "#/definitions/Organization"
                }
            }
        },
...
        "ErrorModel": {
            "type": "object",
            "required": [
                "status",
                "message"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}